+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/miband/MiBandSampleProvider.java
@@ -23,7 +23,7 @@ public class MiBandSampleProvider implements SampleProvider {
     private final float movementDivisor = 180.0f; //256.0f;
 
     @Override
-    public int normalizeType(byte rawType) {
+    public int normalizeType(int rawType) {
         switch (rawType) {
             case TYPE_DEEP_SLEEP:
                 return ActivityKind.TYPE_DEEP_SLEEP;
@@ -42,7 +42,7 @@ public class MiBandSampleProvider implements SampleProvider {
     }
 
     @Override
-    public byte toRawActivityKind(int activityKind) {
+    public int toRawActivityKind(int activityKind) {
         switch (activityKind) {
             case ActivityKind.TYPE_ACTIVITY:
                 return TYPE_ACTIVITY;
@@ -59,12 +59,12 @@ public class MiBandSampleProvider implements SampleProvider {
     }
 
     @Override
-    public float normalizeIntensity(short rawIntensity) {
+    public float normalizeIntensity(int rawIntensity) {
         return rawIntensity / movementDivisor;
     }
 
     @Override
-    public byte getID() {
+    public int getID() {
         return SampleProvider.PROVIDER_MIBAND;
     }
 }
